// ui/src/utils/dashboard_client/api.ts
export interface DiagnosticspbSearchLogRequest {
    /**
     * 
     * @type {number}
     * @memberof DiagnosticspbSearchLogRequest
     */
    end_time?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof DiagnosticspbSearchLogRequest
     */
    levels?: Array<number>;
    /**
     * We use a string array to represent multiple CNF pattern sceniaor like: SELECT * FROM t WHERE c LIKE \'%s%\' and c REGEXP \'.*a.*\' because Golang and Rust don\'t support perl-like (?=re1)(?=re2)
     * @type {Array<string>}
     * @memberof DiagnosticspbSearchLogRequest
     */
    patterns?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof DiagnosticspbSearchLogRequest
     */
    start_time?: number;
}
// ...
export interface LogsearchCreateTaskGroupRequest {
    /**
     * 
     * @type {DiagnosticspbSearchLogRequest}
     * @memberof LogsearchCreateTaskGroupRequest
     */
    request: DiagnosticspbSearchLogRequest;
    /**
     * 
     * @type {Array<LogsearchSearchTarget>}
     * @memberof LogsearchCreateTaskGroupRequest
     */
    search_targets: Array<LogsearchSearchTarget>;
}
// ...
export interface LogsearchSearchTarget {
    /**
     * 
     * @type {string}
     * @memberof LogsearchSearchTarget
     */
    ip?: string;
    /**
     * 
     * @type {number}
     * @memberof LogsearchSearchTarget
     */
    kind?: number;
    /**
     * 
     * @type {number}
     * @memberof LogsearchSearchTarget
     */
    port?: number;
    /**
     * 
     * @type {number}
     * @memberof LogsearchSearchTarget
     */
    status_port?: number;
}